+++ packages/conversation/src/models/v1/app-event/app-event.ts
@@ -1,5 +1,6 @@
 import { ComposingEvent } from '../composing-event';
 import { ComposingEndEvent } from '../composing-end-event';
+import { ReadMessageEvent } from '../read-message-event';
 import { CommentReplyEvent } from '../comment-reply-event';
 import { AgentJoinedEvent } from '../agent-joined-event';
 import { AgentLeftEvent } from '../agent-left-event';
@@ -11,6 +12,7 @@ import { GenericEvent } from '../generic-event';
 export type AppEvent =
   AppComposingEvent
   | AppComposingEndEvent
+  | AppReadMessageEvent
   | AppCommentReplyEvent
   | AppAgentJoinedEvent
   | AppAgentLeftEvent
@@ -19,6 +21,7 @@ export type AppEvent =
 interface AppComposingEvent extends ComposingEvent {
   // Exclude other event types
   composing_end_event?: never;
+  read_message_event?: never;
   comment_reply_event?: never;
   agent_joined_event?: never;
   agent_left_event?: never;
@@ -28,16 +31,28 @@ interface AppComposingEvent extends ComposingEvent {
 interface AppComposingEndEvent extends ComposingEndEvent {
   // Exclude other event types
   composing_event?: never;
+  read_message_event?: never;
   comment_reply_event?: never;
   agent_joined_event?: never;
   agent_left_event?: never;
   generic_event?: never
 }
 
+interface AppReadMessageEvent extends ReadMessageEvent {
+  // Exclude other event types
+  composing_event?: never;
+  composing_end_event?: never;
+  comment_reply_event?: never;
+  agent_joined_event?: never;
+  agent_left_event?: never;
+  generic_event?: never;
+}
+
 interface AppCommentReplyEvent extends CommentReplyEvent {
   // Exclude other event types
   composing_event?: never;
   composing_end_event?: never;
+  read_message_event?: never;
   agent_joined_event?: never;
   agent_left_event?: never;
   generic_event?: never;
@@ -47,6 +62,7 @@ interface AppAgentJoinedEvent extends AgentJoinedEvent {
   // Exclude other event types
   composing_event?: never;
   composing_end_event?: never;
+  read_message_event?: never;
   comment_reply_event?: never;
   agent_left_event?: never;
   generic_event?: never;
@@ -56,6 +72,7 @@ interface AppAgentLeftEvent extends AgentLeftEvent {
   // Exclude other event types
   composing_event?: never;
   composing_end_event?: never;
+  read_message_event?: never;
   comment_reply_event?: never;
   agent_joined_event?: never;
   generic_event?: never;
@@ -65,6 +82,7 @@ interface AppGenericEvent extends GenericEvent {
   // Exclude other event types
   composing_event?: never;
   composing_end_event?: never;
+  read_message_event?: never;
   comment_reply_event?: never;
   agent_joined_event?: never;
   agent_left_event?: never;